New assets for level building and minor changes
The game is literally unplayable for me in this build, so maybe we'll have to revert versions. Gravity scale also did not seem to change, even when I updated its value on the player prefab. It would just be immediately overwritten and reset on game start. I'd like gravity to be a little heavier.